fix tests path
