+++ apps/search/src/test/handlers/category_deleted.test.ts
@@ -1,16 +1,16 @@
 import { createMocks } from "node-mocks-http";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
-import { AlgoliaInvalidAppIdError } from "../../../../lib/algolia/algolia-errors";
-import { createSearchProblemReporter } from "../../../../modules/app-problems";
-import { createWebhookContext } from "../../../../webhooks/webhook-context";
-import { handler } from "./product_updated";
+import { AlgoliaInvalidAppIdError } from "../../lib/algolia/algolia-errors";
+import { createSearchProblemReporter } from "../../modules/app-problems";
+import { handler } from "../../pages/api/webhooks/saleor/category_deleted";
+import { createWebhookContext } from "../../webhooks/webhook-context";
 
-vi.mock("../../../../webhooks/webhook-context", () => ({
+vi.mock("../../webhooks/webhook-context", () => ({
   createWebhookContext: vi.fn(),
 }));
 
-vi.mock("../../../../lib/logger", () => ({
+vi.mock("../../lib/logger", () => ({
   createLogger: () => ({
     info: vi.fn(),
     warn: vi.fn(),
@@ -19,19 +19,17 @@ vi.mock("../../../../lib/logger", () => ({
 }));
 
 const mockReportAuthError = vi.fn();
-const mockReportRecordTooLarge = vi.fn();
 const mockReportInvalidAppIdAndDeactivate = vi.fn();
 
-vi.mock("../../../../modules/app-problems", () => ({
+vi.mock("../../modules/app-problems", () => ({
   createSearchProblemReporter: vi.fn(() => ({
     reportAuthErrorAndDeactivate: mockReportAuthError,
-    reportRecordTooLarge: mockReportRecordTooLarge,
     reportInvalidAppIdAndDeactivate: mockReportInvalidAppIdAndDeactivate,
   })),
 }));
 
 const mockContext = {
-  event: "PRODUCT_UPDATED",
+  event: "CATEGORY_DELETED",
   authData: {
     appId: "app-id",
     domain: "domain.saleor.io",
@@ -39,41 +37,39 @@ const mockContext = {
     saleorApiUrl: "https://domain.saleor.io/graphql/",
   },
   payload: {
-    __typename: "ProductUpdated" as const,
-    product: {
-      id: "prod123",
-      variants: [],
+    __typename: "CategoryDeleted" as const,
+    category: {
+      id: "cat123",
     },
   },
 };
 
-describe("product_updated webhook handler", () => {
+describe("category_deleted webhook handler", () => {
   beforeEach(() => {
     vi.resetAllMocks();
     mockReportAuthError.mockResolvedValue(undefined);
-    mockReportRecordTooLarge.mockResolvedValue(undefined);
     mockReportInvalidAppIdAndDeactivate.mockResolvedValue(undefined);
   });
 
-  it("Returns 200 when no product in payload", async () => {
+  it("Returns 200 when no category in payload", async () => {
     const { req, res } = createMocks();
 
-    const contextWithoutProduct = {
+    const contextWithoutCategory = {
       ...mockContext,
-      payload: { __typename: "ProductUpdated" as const, product: null },
+      payload: { __typename: "CategoryDeleted" as const, category: null },
     };
 
     // @ts-expect-error - mocking request for testing
-    await handler(req, res, contextWithoutProduct);
+    await handler(req, res, contextWithoutCategory);
 
     expect(res._getStatusCode()).toBe(200);
   });
 
-  it("Returns 200 on successful update", async () => {
+  it("Returns 200 on successful delete", async () => {
     const { req, res } = createMocks();
 
     const mockAlgoliaClient = {
-      updateProduct: vi.fn().mockResolvedValue(undefined),
+      deleteCategory: vi.fn().mockResolvedValue(undefined),
     };
 
     vi.mocked(createWebhookContext).mockResolvedValue({
@@ -84,37 +80,7 @@ describe("product_updated webhook handler", () => {
     await handler(req, res, mockContext);
 
     expect(res._getStatusCode()).toBe(200);
-    expect(mockAlgoliaClient.updateProduct).toHaveBeenCalledWith(mockContext.payload.product);
-  });
-
-  it("Returns 413 with actionable error message when Algolia record size limit exceeded", async () => {
-    const { req, res } = createMocks();
-
-    const recordSizeError = {
-      status: 400,
-      message: "Record at the position 0 objectID=prod123_var456 is too big size=15000/10000 bytes",
-    };
-
-    const mockAlgoliaClient = {
-      updateProduct: vi.fn().mockRejectedValue(recordSizeError),
-    };
-
-    vi.mocked(createWebhookContext).mockResolvedValue({
-      algoliaClient: mockAlgoliaClient,
-    } as any);
-
-    // @ts-expect-error - mocking request for testing
-    await handler(req, res, mockContext);
-
-    expect(res._getStatusCode()).toBe(413);
-
-    const responseBody = res._getData();
-
-    expect(responseBody).toContain("exceeds Algolia's record size limit");
-    expect(responseBody).toContain("Current size: 15000 bytes");
-    expect(responseBody).toContain("Algolia fields filtering");
-    expect(responseBody).toContain("algoliaDescription");
-    expect(responseBody).toContain("https://docs.saleor.io");
+    expect(mockAlgoliaClient.deleteCategory).toHaveBeenCalledWith("cat123");
   });
 
   it("Returns 500 on other Algolia errors", async () => {
@@ -123,7 +89,7 @@ describe("product_updated webhook handler", () => {
     const otherError = new Error("Some other error");
 
     const mockAlgoliaClient = {
-      updateProduct: vi.fn().mockRejectedValue(otherError),
+      deleteCategory: vi.fn().mockRejectedValue(otherError),
     };
 
     vi.mocked(createWebhookContext).mockResolvedValue({
@@ -155,7 +121,7 @@ describe("product_updated webhook handler", () => {
     const authError = { status: 403, message: "Invalid Application-ID or API key" };
 
     const mockAlgoliaClient = {
-      updateProduct: vi.fn().mockRejectedValue(authError),
+      deleteCategory: vi.fn().mockRejectedValue(authError),
     };
 
     vi.mocked(createWebhookContext).mockResolvedValue({
@@ -170,43 +136,15 @@ describe("product_updated webhook handler", () => {
     expect(vi.mocked(createSearchProblemReporter)).toHaveBeenCalledWith(mockContext.authData);
   });
 
-  it("Reports record too large problem when Algolia record size limit exceeded", async () => {
-    const { req, res } = createMocks();
-
-    const recordSizeError = {
-      status: 400,
-      message: "Record at the position 0 objectID=prod123_var456 is too big size=15000/10000 bytes",
-    };
-
-    const mockAlgoliaClient = {
-      updateProduct: vi.fn().mockRejectedValue(recordSizeError),
-    };
-
-    vi.mocked(createWebhookContext).mockResolvedValue({
-      algoliaClient: mockAlgoliaClient,
-    } as any);
-
-    // @ts-expect-error - mocking request for testing
-    await handler(req, res, mockContext);
-
-    expect(res._getStatusCode()).toBe(413);
-    expect(mockReportRecordTooLarge).toHaveBeenCalledWith({
-      type: "product_variant",
-      productId: "prod123",
-      variantId: "var456",
-    });
-    expect(vi.mocked(createSearchProblemReporter)).toHaveBeenCalledWith(mockContext.authData);
-  });
-
   it("Returns 401 and deactivates app when Algolia Application ID is invalid", async () => {
     const { req, res } = createMocks();
 
-    const invalidAppIdError = new AlgoliaInvalidAppIdError(
-      "Algolia Application ID does not exist or is unreachable",
-    );
-
     const mockAlgoliaClient = {
-      updateProduct: vi.fn().mockRejectedValue(invalidAppIdError),
+      deleteCategory: vi
+        .fn()
+        .mockRejectedValue(
+          new AlgoliaInvalidAppIdError("Algolia Application ID does not exist or is unreachable"),
+        ),
     };
 
     vi.mocked(createWebhookContext).mockResolvedValue({
@@ -219,6 +157,5 @@ describe("product_updated webhook handler", () => {
     expect(res._getStatusCode()).toBe(401);
     expect(res._getData()).toContain("Algolia Application ID does not exist");
     expect(mockReportInvalidAppIdAndDeactivate).toHaveBeenCalledWith("app-id");
-    expect(vi.mocked(createSearchProblemReporter)).toHaveBeenCalledWith(mockContext.authData);
   });
 });
