+++ apps/search/src/test/handlers/product_variant_updated.test.ts
